requirements.txt
# Master requirements.txt for Tickles And Co
# Consolidated from all project requirements files

# --- Core Data Processing ---
numpy>=2.0.0
pandas>=2.2.0

# --- Technical Analysis ---
TA-Lib>=0.4.28

# --- Database ---
mysql-connector-python>=8.0.0
pymysql>=1.1.0
cryptography>=43.0.0
dbutils>=3.1.0

# --- Environment & System ---
python-dotenv>=1.0.0
psutil>=5.9.0
requests>=2.31.0

# --- Performance Optimization ---
numba>=0.57.0
llvmlite>=0.40.0

# --- Web Frameworks ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
flask>=3.1.0
jinja2>=3.1.0

# --- AI Model APIs ---
anthropic>=0.40.0
openai>=1.55.0
google-generativeai>=0.8.0

# --- Vector Database ---
qdrant-client>=1.12.0

# --- Memory Layer ---
mem0ai>=0.1.0

# --- Local Embeddings ---
sentence-transformers>=2.2.0

# --- Cloud Vector Store (Optional) ---
pinecone-client>=3.0.0

# --- Exchange Connectivity ---
ccxt>=4.0.0

# --- News & RSS ---
feedparser>=6.0.0
aiohttp>=3.10.0

# --- Headless Browser ---
playwright>=1.49.0

# --- Discord ---
discord.py-self>=2.0.0
curl_cffi>=0.15.0
PyNaCl>=1.5.0

# --- Telegram ---
telethon>=1.37.0

# --- Web Automation ---
selenium>=4.18.0
webdriver-manager>=4.0.0

# --- Charting & Visualization ---
plotly>=5.24.0

# --- Utilities ---
pydantic>=2.9.0
httpx>=0.27.0
apscheduler>=3.10.0

# --- Market Data ---
yfinance>=0.2.37